# Conflicts:
#	helm/hpcc/Chart.yaml
#	helm/hpcc/templates/_helpers.tpl
#	helm/hpcc/templates/dafilesrv.yaml
#	helm/hpcc/templates/dali.yaml
#	helm/hpcc/templates/dfuserver.yaml
#	helm/hpcc/templates/eclagent.yaml
#	helm/hpcc/templates/eclccserver.yaml
#	helm/hpcc/templates/eclscheduler.yaml
#	helm/hpcc/templates/esp.yaml
#	helm/hpcc/templates/localroxie.yaml
#	helm/hpcc/templates/roxie.yaml
#	helm/hpcc/templates/sasha.yaml
#	helm/hpcc/templates/thor.yaml
#	version.cmake

+++ system/jlib/jptree.cpp
@@ -6607,7 +6607,7 @@ void printJSON(const IPropertyTree *tree, unsigned indent, byte flags)
     printf("%s", json.str());
 }
 
-void dbglogJSON(const IPropertyTree *tree, unsigned indent, unsigned flags)
+void dbglogJSON(const IPropertyTree *tree, unsigned indent, byte flags)
 {
     StringBuffer json;
     toJSON(tree, json, indent, flags);
